batch_scheduler/requirements.txt
Flask==1.1.2
psycopg2-binary==2.9.9
Werkzeug==2.0.3
jinja2==3.0.2
flask_login==0.5.0
flask_migrate==2.6.0
WTForms==3.0.0
flask_wtf==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-Cors==4.0.1
flask-socketio==5.3.6
sqlalchemy==1.4.29
email_validator==1.1.3
gunicorn==20.1.0
flask-restx==0.5.1
python-dotenv==1.0.1
Flask-Minify==0.37
Flask-Script==2.0.6
python-logging-loki==0.3.1
python-dateutil==2.9.0
scp==0.15.0
itsdangerous==2.0.1
PyYAML==6.0.1
concurrent-log-handler==0.9.25
ngram==4.0.3
pytest==8.3.2
coverage==7.6.0
PyJWT==2.8.0
cryptography==43.0.1
redis
kombu
pygtail==0.14.0
opentelemetry-exporter-otlp==1.34.1
opentelemetry-instrumentation-flask==0.55b1
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-psycopg2>=0.55b1
opentelemetry-instrumentation-redis>=0.55b1
opentelemetry-instrumentation-sqlite3>=0.55b1
opentelemetry-instrumentation-urllib>=0.55b1
opentelemetry-instrumentation-urllib3>=0.55b1
pydantic==2.11.7
